requirements.txt
Django==5.2.11
Faker==40.5.1
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3
requests==2.32.3
python-dotenv==1.0.1
django-allauth==65.15.1
PyJWT==2.12.1
cryptography==46.0.6
